UserProfileViewModelをインタフェースと実装に分ける
