+++ app/src/main/java/com/example/foos/ui/view/screen/userprofile/UserProfileScreen.kt
@@ -33,7 +33,7 @@ import moe.tlaster.nestedscrollview.rememberNestedScrollViewState
 @OptIn(ExperimentalPagerApi::class)
 @Composable
 fun UserProfileScreen(
-    viewModel: UserProfileViewModel,
+    viewModel: UserProfileViewModelImpl,
     navController: NavController,
     userId: String
 ) {
